+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - add OpenID Connect RP-Initiated Logout 1.0 support
 - allow specify the default Magda orgUnit id that the user should be assigned to when logged in for the first time
+- #8: allow specify the default Magda role id that the user should be granted when logged in for the first time
 
 # v2.0.0